pyproject.toml
[tool.poetry]
name = "friends-wallet"
version = "0.1.0"
description = ""
authors = ["Timur Alchin"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.*"
Django = "^4.1.3"
django-colorfield = "^0.8.0"
djangorestframework = "^3.14.0"
djangorestframework-jwt = "^1.11.0"
drf-yasg = "^1.21.5"
djangorestframework-simplejwt = {extras = ["crypto"], version = "^5.2.2"}
django-stubs = "^4.2.0"
django-split-settings = "^1.2.0"
python-decouple = "^3.8"
factory-boy = "3.2.1"
pytest-env = "^0.8.1"
wemake-python-styleguide = "^0.18.0"
pytest-django = "4.5.2"
pytest = "7.2.1"
django-cors-headers = "^4.0.0"
psycopg2-binary = "^2.9.6"
mock = "^5.0.2"


[tool.poetry.dev-dependencies]
ipython = "^8.13.2"
isort = "^5.12.0"
autopep8 = "^2.0.2"
django-debug-toolbar = "^4.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
